feat(options): Add default editor options
This being the first commit I also started a `CHANGELOG` and a `README`.
The scope is not to duplicate the same info from the git log into the
changelog.

The scope of the changelog is to give a clear overview of what was
added, changed, fixed or broken between version. Indeed conventional
commits should help having a good changelog and generating correct
semver versions.

Let me put it in another way, the changelog is for everyday humans, the
git log are for devs. This might sound harsh for some, but deal with it,
different material has different audiences and that's OK.