feat(telescope): Add `telescope` plugin with usefull keybindings
`telescope` is a really awesome plugin. It helps searching strings in
files, files in current working directory, keybindings  and other
sources. Most of the time it eliminates the need for `netrw` usage, it
does not eliminate it. When there is a need to see a global files
hierarchy `netrw` is the way to do.

The following keybindings were configures:

- `<LEADER>ff` - Find files in current working directory
- `<LEADER>fs` - Find string in current working directory
- `<LEADER>fc` - Find string under cursor in current working directory
- `<LEADER>fb` - Fuzzy find open buffers
- `<LEADER>fk` - Fuzzy find keybindings
- In `telescope`, `Ctrl+J` - Go to next selection
- In `telescope`, `Ctrl+K` - Go to previous selection

There are more features exposed by `telescope` directly or through other
plugins, but for now those settings are enough.