fix(lazy): Use `vim.uv` or `vim.loop` as alternative
When using `neovim` `0.9.x` we have access to `vim.loop`, but with
`neovim` `0.10.x` `vim.uv` was introduced.

This change introduces support for both. First it tries to use `vim.uv`,
if it is not available it will use `vim.loop`.

In this way we make sure we support both.