feat(nvim-cmp): Add completion for words used in current buffers
For the completion popup we use `nvim-cmp` plugin and as source for the
words in current buffers we use `cmp-buffer` plugin.

This kind of completion is usefull when working on documentation and
some words you don't really remember and you can look up after them or
you can use it as a way to type faster.