fix(telescope): Include hidden files for `find_files` command
The default behavior is wrong. It's helpful to search into hidden files.
A good example is when you have hidden folders in your project and you
need to interact with them.

Good examples: `.github`.

By allowing the files finder to look also for hidden files then we will
able to find a GitHub workflow file using the fuzzy finder.