fix(telescope): Remove `smart` path display in `telescope`
For the time being this is really confusing and I don't get the smart
part of it.

For me the biggest confusion was when for a files hierarchy like this
one:

```
- project
  - lua
    - plugins
      - plugin1.lua
      - plugin2.lua
```

In `telescope` the paths were shown with something like:

```
../plugins/plugin1.lua
```

The provided example is not the exact problem I faced, but it is in it's
spirit. It was super confusing to see that `../` when I knew the file
was somewhere down the hierarchy, not up.

Maybe I'll reintroduce it in the future, but for the moment this stays
out of `typewriter`.