revert: Merge branch `plugin_layers`
The split into configuration layers added much unneeded complexity at
elast for me. Because of this I decided to work on a single
configuration based on `Lazy.nvim` and when everything works, it is
tested, only then, maybe, just maybe I will take the axe to it and start
modularizing it.

For me `LazyVim` is a great source of inspiration, but it still looks
like magic to me so I woun't do thinkg `LazyVim` way for now.

Yes I'm not using `LazyVim` distribution, but I just get inspired by it.
I want to be aware of everything that goes into my own `neovim`
configuration.

This commit reverts the merge done in `2aa5440a`