feat(lspconfig): Add default LSP configuration
With this commit the diagnostics icons of the LSP are customized to look
better.

Also we set keybindings for a few basic LSP functionality using an auto
command that triggeres when an LSP attaches to the buffer.

- Normal mdoe, `gr`: Show all references in `telescope`

- Normal mode, `gd`: Go to definition or list all options in
  `telescope`

- Normal mode `gi`: Go to implementation or list all options in
  `telescope`

- Normal mode, `rs`: Restart LSP server

For now there is only a default LSP handler for all LSPs, there is no
custom setup.