feat(dap): Add support for attaching to C# processes
Whent the "Attach to process' DAP configuration is chosen a `telescope`
picker will show up with all the running processes listed. You can use
this picker to search for the process you want to attach to.

Keep in mind that if a process was started after the picker was shown,
that new process will not be shown in the list of processes.

For now there is no plan to support auto refreshing of the process list.