feat(lsp): Format the current buffer
For now we format the buffer using the LSP. It is good enough for now.

It's work looking if LSP formatting is enough and if it supports
`.editorconfig` files or if we need to use dedicated formatters per
language.

As for the keybinding we are using my long time favorite `<LEADER>f`.