feat(dap): Add suport for launching a DLL and debug it
First we run a `dotnet build -c Debug` to build the project in `debug`
mode.

Then we are looking for all DLL files and present a `telescope` picker
to choose the DLL we want to launch and debug.

For now this is enough. Ideally we should launch the project and attach
to it automatically. For solutions we should launch the first project of
the solution. Anyway this is a battle for another day.