feat(dap): Add support for launching and debug Rust binaries
As debug adapter I use `codelldb`. It is installed using `mason`.

The DAP UI is different from C#. It's nice that we get access to the CPU
registers, but the UI seems a bit clutered when debugging Rust code.

It's worth considering improving the DAP UI setu in the future.