chore(TODO): Add a TODO file to track ideas and stuff todo
As I work alone I don't want to go trough GitHub issues or a Trello
board to manage the tasks I'm working on. I believe that for time being
this file is all I need.

The `Done` section might overlap with the entries in `CHANGELOG`, but it
will contain also refactorings and thing that are irrelevant for the
final user.