test: Run test containers as `root`, not `test` user.
Because I want to transition to a dev containers inspired workflow I
need to be able to bind mount, read and write, the project directory
inside the container.

When running as `test` user if you bind mound `/host/dir` you can read
the data, but you can't write to the mounted directory. You can go
around it by doing some fiddling with the access rights on the host, but
I dont' want to do that I just want to be able to spin the test
container, test `typewriter` and if needed use `typewriter` from inside
the test container to adjust the code of `typewriter` and those changes
will be visible on the host as well.

Regarding security, the `root` user fromt he container is not the `root`
user from the host. Also these containers are meant for testing during
development. They are not meant for production where unknown actors can
interact with it remotely, like it would be for a deployed web service.

This commit does not implement all that, I still copy `typewriter`
inside the container. With this intermediary step I want to make sure
running the test containers as `root` works well.