test: Bind mount `typewriter` in test containers
With this commit we bind mount `typewriter` in the test containers.

This two way code change between host and test container is useful
because sometimes you notice a bug and you want to fix it.

With the old worflow where a new `typewriter` container image was
created each time the code changed it was cumbersome to always build and
run the image for every code change. I also found myself edititing the
code from the image, but forgetting to apply the fixes in the code from
the host.

With this workflow when I fix a bug inside the container the changes
will be visible on the host and after I validate the fix I can go ahead
and commit them from the host. This also eliminates the need to
continuously build new `typewriter` test images. This will also reduce
plution with dangling container images.