refactor(dap): Remove duplicated launch target picker code
To be able to use a common scaffolding for the launch target picker we
need to build the binary differently for different platforms, also
looking for the generated binary differes between languages and
platforms.

For now `typewriter` supports DAP for `dotnet` and `rust`, those use
different commands for building. The generated binary is a DLL for
`dotnet` and for `rust` is a classic executable. The build output
directory is different etc.

To cope with different build command and binary finders I decided to
crate two maps, `build_cmd` and `finder_cmd`.

```lua
local build_cmd = {
    cs = "dotnet build -c Debug",
}

local finder_cmd = {
    cs = { "find", ".", "-type", "f", "*.dll" },
}
```

`build_cmd` is a string, it will be passed to `os.execute()`.

`finder_cmd` is an array of strings. The requirement of defining it like
this comes from the fact that behind the scenes I pass it to
`finders.new_oneshot_job()`. Instad of daling with splitting a string
into it's parts I decided to define the `finder_cmd` as required by
`new_oneshot_job()`.

From the example above `cs` is the `vim.bo.filetype` of the current
buffer. I'm not sure this is the best way to detect which builder and
finder to use, but it's a good start.

I'm curious how this setup handles multi project workspaces. Time will
tell.