fix(vim-tmux-nav): Fix navigation in `netrw` split
By default `Ctrl+L` is used by `netrw` to execute `NetrwRefresh`
command. To make sure `netrw` is not interferring with splits navigation
I decided to overwrite `Ctrl+L` for `netrw` as well.

This will create a discrepancy between `netrw`'s help manual and the
reality when it comes to the keybindings defined in `typewriter`, but it
is a needed evil. In my view easy split navigation is more important
than refreshing `netrw`.