refactor(tests): Merge `containerfiles` into `tests`
There is no reason to keep `containerfiles` and `tests` as a separate
module, but it makes sense to have `containers` as a submodule of
`tests`. By doing so we focus changes related to test containers in a
single place instead of running around and updating two separate
modules.

For the end user this introduces no braking changes, but for developers
and testers this introduces a braking change.

From now on we should use:

- `./tests/containers/fedora_41`
- `./tests/containers/ubuntu_24.04`

Instead of:

- `./tests/test_fedora_41`
` `./tests/test_ubuntu_24.01`