refactor: Rename `me` module to `typewriter`
The rationale behind this renaming comes from the fact that it's better
to place the implementation of `typewriter` in the `typewriter` module
rather than having it in the `me` module.

When you look at both names `typewriter` clearly says what the modules
is all about, while `me` is just vague and can contain anything.