feat(nvim-tree): Open file tree in floating window
When starting Neovim with `nvim dir/`, then `nvim-tree` will occupy the
entire screen. When you open a file with `nvim file`, the file will show
up. Now when activating `nvim-tree` with `:NvimTreeToggle` it will be
shown in a floating Neovim window that is centered and occupying 80% of
the editor window.